--- a/src/llmcompressor/args/model_arguments.py
+++ b/src/llmcompressor/args/model_arguments.py
@@ -81,7 +81,7 @@ class ModelArguments:
         metadata={"help": "Whether to compress sparse models during save"},
     )
     oneshot_device: Optional[str] = field(
-        default="cuda:0",
+        default="cuda",
         metadata={"help": "Device to run oneshot calibration on"},
     )
     model_revision: str = field(