--- a/src/llmcompressor/entrypoints/oneshot.py
+++ b/src/llmcompressor/entrypoints/oneshot.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 import torch
 from accelerate.hooks import remove_hook_from_module
-from compressed_tensors.utils import force_cpu_offload
+from compressed_tensors.utils import offloaded_dispatch
 from loguru import logger
 from torch.utils.data import DataLoader
 from transformers import PreTrainedModel
@@ -129,7 +129,7 @@ class Oneshot:
         # offload to cpu if possible
         if "cuda" in str(model_args.oneshot_device) and torch.cuda.is_available():
             remove_hook_from_module(model_args.model, recurse=True)
-            force_cpu_offload(model_args.model, model_args.oneshot_device)
+            offloaded_dispatch(model_args.model, model_args.oneshot_device)
         else:
             logger.warning("CUDA is not available! Compressing model on CPU instead")