--- a/src/llmcompressor/pipelines/basic/pipeline.py
+++ b/src/llmcompressor/pipelines/basic/pipeline.py
@@ -9,6 +9,7 @@ from llmcompressor.core import LifecycleCallbacks
 from llmcompressor.modifiers.utils.pytorch_helpers import apply_pad_mask_to_batch
 from llmcompressor.pipelines.registry import CalibrationPipeline
 from llmcompressor.pytorch.utils.helpers import tensors_to_device
+from llmcompressor.utils.dev import dispatch_for_generation
 from llmcompressor.utils.helpers import calibration_forward_context
 
 if TYPE_CHECKING:
@@ -37,6 +38,7 @@ class BasicPipeline(CalibrationPipeline):
         :param dataloader: loads data for calibration
         :param dataset_args: dataset arguments relevant to pipelines
         """
+        dispatch_for_generation(model)
         model_device = get_execution_device(model)
 
         LifecycleCallbacks.calibration_epoch_start()