--- a/src/llmcompressor/pipelines/layer_sequential/pipeline.py
+++ b/src/llmcompressor/pipelines/layer_sequential/pipeline.py
@@ -16,8 +16,8 @@ from llmcompressor.pipelines.layer_sequential.helpers import (
 )
 from llmcompressor.pipelines.registry import CalibrationPipeline
 from llmcompressor.pipelines.sequential.helpers import (
+    disable_offloading,
     get_targets_from_modifiers,
-    keep_onload_context,
 )
 from llmcompressor.utils.helpers import DisableQuantization, calibration_forward_context
 
@@ -88,7 +88,7 @@ class LayerSequentialPipeline(CalibrationPipeline):
                 prop_desc = f"({layer_index + 1}/{num_layers}): Propagating"
 
                 # reduce memory movement by keeping modules onloaded
-                with keep_onload_context():
+                with disable_offloading():
                     # do a preliminary pass to trigger modifier hooks
                     for batch_idx in tqdm.tqdm(range(len(dataloader)), desc=calib_desc):
                         inputs = intermediates.fetch(batch_idx)