@kylesayrs 2025-06-10
Spoke with @brian-dellabetta, we imagine an optimal example script might look something like this

```python3
model = load(..., device_map="cpu")
model = oneshot(model)
with compressed_tensors.compress(model):
    model.save_pretrained(...)

dispatch_model(model, device_map="auto")
... run sample generation
```

```python3
model = load(..., device_map="cpu")
model = oneshot(model)
compressed_tensors.compress(model)
model.save_pretrained(...)

compressed_tensors.decompress(model)
dispatch_model(model, device_map="auto")
... run sample generation
```