@logos-mhr 2025-06-11
Excellent job! Sequential onloading is definitly essential for larger models.
But I encountered a potential bug when trying to apply GPTQ to DeepSeek v2-lite, which made` weight_scale` and `weight_zero_point` of the quantized Linear modules meta tensors.
It seems this may be caused by the hook of  `force_cpu_offload`.  When applying that hook, the Linear modules does not have ` weight_scale` and `weight_zero_point` attributes. Could that be the potiential cause? Or could it be solved by simply adding place holding tensors?