@logos-mhr 2025-06-12
> Excellent job! Sequential onloading is definitly essential for larger models. But I encountered a potential bug when trying to apply GPTQ to DeepSeek v2-lite, which made` weight_scale` and `weight_zero_point` of the quantized Linear modules meta tensors. It seems this may be caused by the hook of `force_cpu_offload`. When applying that hook, the Linear modules does not have ` weight_scale` and `weight_zero_point` attributes. Could that be the potiential cause? Or could it be solved by simply adding place holding tensors?

Apply `remove_hook_from_module` and `force_cpu_offload`  after `session.initialize` in `one_shot.py` will do the job. This may indicate my assumption above is correct.