@kylesayrs 2025-06-12
Hi @logos-mhr, I was able to replicate the issue I believe you're referring to, namely that the qparams are on the meta device after the hooks is removed. I'll put up a fix soon