@lyzhongcrd 2025-06-16
Great job! I tried to quantize Qwen2.5-VL-32B on an A100 (failed due to OOM) and found this PR. When shall it be released? Can't wait to use this new feature.